+++ mkdocs_pdf_generate/styles/__init__.py
@@ -56,11 +56,11 @@ def style_for_print(options: Options, pdf_metadata: Dict) -> List[Tag]:
         _css_escape(options.author),
         _css_escape(options.author_logo),
         _css_escape(options.copyright),
-        _css_escape(pdf_metadata.get("title", options.body_title or options.cover_title)),
-        _css_escape(pdf_metadata.get("subtitle", options.cover_subtitle)),
-        _css_escape(pdf_metadata.get("type", "Documentation")),
-        _css_escape(pdf_metadata.get("revision", "")),
-        _css_escape(pdf_metadata.get("filename", "")),
+        _css_escape(pdf_metadata.get("title") or options.body_title or options.cover_title),
+        _css_escape(pdf_metadata.get("subtitle") or options.cover_subtitle),
+        _css_escape(pdf_metadata.get("type") or "Documentation"),
+        _css_escape(pdf_metadata.get("revision") or ""),
+        _css_escape(pdf_metadata.get("filename") or ""),
         _css_escape(re.sub(r"http://|https://", "", options.site_url)),
     )
     css_tag = Tag(name="style", attrs={"class": "plugin-default-css"})